crypto/witches_symmetric_exam/solver/requirements.txt
-i https://pypi.org/simple
certifi==2022.9.24 ; python_version >= '3.6'
chardet==3.0.4
idna==2.8
pycryptodome==3.15.0
requests==2.22.0
urllib3==1.25.11 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
